edit

+++ pages/CRUD/delete_group.php
@@ -2,6 +2,10 @@
 $group_id = isset($_GET['group']) ? $_GET['group'] : '';
 
 
+  $sql = "DELETE FROM course_group_rel WHERE group_id = $group_id";
+if (!mysqli_query($conn, $sql)) {
+    echo "Error updating record: " . mysqli_error($conn);
+  }
   $sql = "DELETE FROM group_teacher_rel WHERE group_id = $group_id";
 if (!mysqli_query($conn, $sql)) {
     echo "Error updating record: " . mysqli_error($conn);
@@ -14,5 +18,5 @@ $sql = "DELETE FROM groups WHERE group_id = $group_id";
 if (!mysqli_query($conn, $sql)) {
     echo "Error updating record: " . mysqli_error($conn);
   }
-  header('location:../administrator_page/group_form.php')
+  header('location:' . BASE_URL . '?mode=view&page=groups');
   ?>
